# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agorich Pharma API
  description: |
    Comprehensive API documentation for Agorich Pharma's hyperlocal pharma intelligence operating system.

    ## Overview
    This API provides endpoints for managing the pharmaceutical distribution ecosystem including:
    - AI-powered collection calls via VAPI
    - Inventory management and alerts
    - Invoice processing and approval workflows
    - Real-time metrics and analytics
    - Emergency controls and system management

    ## Authentication
    - **User Authentication**: Supabase Auth with JWT tokens (Cookie-based for web clients)
    - **Agent API Key**: `x-agent-api-key` header for backend-to-backend communication
    - **Admin API Key**: `x-admin-api-key` header for admin operations

    ## Rate Limiting
    - Default: 100 requests per minute per IP
    - AI endpoints: 20 requests per minute
    - Webhook endpoints: 200 requests per minute
  version: 1.0.0
  contact:
    name: Agorich Pharma Tech Team
    email: tech@agorichpharma.com
  license:
    name: Proprietary
servers:
  - url: https://api.agorichpharma.com
    description: Production server
  - url: https://staging-api.agorichpharma.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/agent-connect/apply-discount:
    post:
      tags:
        - Agent Connect
      summary: Apply discount to invoice
      description: |
        Applies a discount to an existing invoice. Requires approval for discounts exceeding thresholds.

        Discount types:
        - percentage: Percentage discount (0-100)
        - fixed: Fixed amount discount in INR
      operationId: applyDiscount
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyDiscountRequest'
            examples:
              percentage:
                summary: Apply 10% discount
                value:
                  invoice_id: "inv-123"
                  discount_type: "percentage"
                  discount_value: 10
                  reason: "Loyalty discount for bulk order"
              fixed:
                summary: Apply fixed discount
                value:
                  invoice_id: "inv-123"
                  discount_type: "fixed"
                  discount_value: 5000
                  reason: "Festival offer"
      responses:
        '200':
          description: Discount applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyDiscountResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Discount exceeds approval threshold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequiredResponse'
...
  /api/approvals/queue:
    get:
      tags:
        - Approvals
      summary: Get approval queue
      description: Returns pending approval requests.
      operationId: getApprovalQueue
      security:
        - BearerAuth: []
      parameters:
        - name: actionType
          in: query
          schema:
            type: string
          description: Filter by action type
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
            default: pending
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Approval queue retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalQueueResponse'
    post:
      tags:
        - Approvals
      summary: Submit for approval
      description: |
        Submits an action for approval. If the action is below the approval threshold,
        it is executed directly without requiring approval.
      operationId: submitForApproval
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-agent-api-key
      description: API key for agent-to-agent communication
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for administrative operations
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
    WebhookSecretAuth:
      type: apiKey
      in: header
      name: x-vapi-webhook-secret
      description: Webhook verification secret
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message description"
      required:
        - success
        - error
...
    ApplyDiscountRequest:
      type: object
      required:
        - invoice_id
        - discount_type
        - discount_value
      properties:
        invoice_id:
          type: string
        discount_type:
          type: string
          enum: [percentage, fixed]
        discount_value:
          type: number
          minimum: 0
        reason:
          type: string
    ApplyDiscountResponse:
      type: object
      properties:
        success:
          type: boolean
        invoice_id:
          type: string
        discount_applied:
          type: number
        new_grand_total:
          type: number
    ApprovalRequiredResponse:
      type: object
      properties:
        success:
          type: boolean
        requiresApproval:
          type: boolean
          example: true
        approvalId:
          type: string
        thresholdInfo:
          type: object
          properties:
            type:
              type: string
            exceededAmount:
              type: number
            message:
              type: string
...
    ApprovalRequest:
      type: object
      required:
        - actionType
        - actionData
      properties:
        actionType:
          type: string
          enum: [discount_exceed, price_override, write_off, user_role_change, emergency_stop]
        actionData:
          type: object
        requestedBy:
          type: string
    ApprovalResponse:
      type: object
      properties:
        success:
          type: boolean
        requiresApproval:
          type: boolean
        approvalId:
          type: string
          nullable: true
        message:
          type: string
...
    ApprovalQueueResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
        count:
          type: integer
...